feat(chart): interactive sale-history chart with chartistry, chips, and stats strip
Replace the PriceHistoryChart stub with a real multi-series scatter chart using
leptos-chartistry 0.2. Adds time-range chips (24h/7d/30d/All), IQR outlier-
filter toggle, and a reactive stats strip (n sales, VWAP, median, min, max).
Also fix two pre-existing clippy::unnecessary_cast errors in gil.rs and
virtual_scroller.rs that were blocking CI.

+++ ultros-frontend/ultros-app/src/components/virtual_scroller.rs
@@ -155,7 +155,7 @@ where
             {
                 // approximate top of target row using measured prefix sums
                 let row_top = target as f64 * row_height + fenwick.with(|f| f.sum(target));
-                let current = div.scroll_top() as f64;
+                let current = div.scroll_top();
                 let visible_top = current + header_h;
                 let visible_bottom = current + header_h + effective_viewport;
                 let row_bottom = row_top + avg_row_height();
